# airbyte-api/src/main/openapi/api.yaml
openapi: 3.0.0
info:
  title: Airbyte Public API
  version: 1.0.0
servers:
  - url: "http://localhost:8006/v1"
    description: Airbyte API v1
paths:
...
  /jobs/{jobId}:
    get:
      tags:
        - Jobs
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
              examples:
                Job Get Response Example:
                  value:
                    id: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    status: running
                    jobType: sync
          description: Get a Job by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: getJob
      summary: Get Job status and details
    delete:
      tags:
        - Jobs
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
          description: Cancel a Job.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: cancelJob
      summary: Cancel a running Job
    parameters:
      - name: jobId
        schema:
          format: int64
          type: integer
        in: path
        required: true
  /sources:
    get:
      tags:
        - Sources
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcesResponse"
          description: Successful operation
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: listSources
      summary: List sources
      parameters:
        - name: workspaceIds
          description: The UUIDs of the workspaces you wish to list sources for. Empty list will retrieve all allowed workspaces.
          schema:
            type: array
            items:
              format: uuid
              type: string
          example: df08f6b0-b364-4cc1-9b3f-96f5d2fccfb2,b0796797-de23-4fc7-a5e2-7e131314718c
          in: query
          required: false
        - name: includeDeleted
          description: Include deleted sources in the returned results.
          schema:
            default: false
            type: boolean
          in: query
          required: false
        - name: limit
          description: Set the limit on the number of sources returned. The default is 20.
          schema:
            format: int32
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          in: query
        - name: offset
          description: Set the offset to start at when returning sources. The default is 0
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          in: query
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceCreateRequest"
            examples:
              Source Creation Request Example:
                value:
                  configuration:
                    airbyte_source_name: google-ads
                    conversion_window_days: 14
                    customer_id: "1234567890"
                    start_date: 2023-01-01
                    end_date: 2024-01-01
                  name: My Source
                  workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
      tags:
        - Sources
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
              examples:
                Source Creation Response Example:
                  value:
                    sourceId: 0c31738c-0b2d-4887-b506-e2cd1c39cc35
          description: Successful operation
        "400":
          description: Invalid data
        "403":
          description: Not allowed
      operationId: createSource
      summary: Create a source
      description: "Creates a source given a name, workspace id, and a json blob containing the configuration for the source."
  /sources/{sourceId}:
    get:
      tags:
        - Sources
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
              examples:
                Source Get Response Example:
                  value:
                    sourceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: running
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
          description: Get a Source by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: getSource
      summary: Get Source details
    patch:
      tags:
        - Sources
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourcePatchRequest"
            examples:
              Source Update Request Example:
                value:
                  configuration:
                    airbyte_source_name: google-ads
                    conversion_window_days: 14
                    customer_id: "1234567890"
                    start_date: 2023-01-01
                    end_date: 2024-01-01
                  name: My Source
                  workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
              examples:
                Source Update Response Example:
                  value:
                    sourceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: running
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
          description: Update a Source
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: patchSource
      summary: Update a Source
    put:
      tags:
        - Sources
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourcePutRequest"
            examples:
              Source Update Request Example:
                value:
                  configuration:
                    airbyte_source_name: google-ads
                    conversion_window_days: 14
                    customer_id: "1234567890"
                    start_date: 2023-01-01
                    end_date: 2024-01-01
                  name: My Source
                  workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
              examples:
                Source Update Response Example:
                  value:
                    sourceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: running
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
          description: Update a source and fully overwrite it
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: putSource
      summary: Update a Source and fully overwrite it
    delete:
      tags:
        - Sources
      responses:
        "204":
          description: The resource was deleted successfully
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: deleteSource
      summary: Delete a Source
    parameters:
      - name: sourceId
        schema:
          format: UUID
          type: string
        in: path
        required: true
  /destinations:
    get:
      tags:
        - Destinations
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationsResponse"
          description: Successful operation
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: listDestinations
      summary: List destinations
      parameters:
        - name: workspaceIds
          description: The UUIDs of the workspaces you wish to list destinations for. Empty list will retrieve all allowed workspaces.
          schema:
            type: array
            items:
              format: uuid
              type: string
          in: query
          required: false
        - name: includeDeleted
          description: Include deleted destinations in the returned results.
          schema:
            default: false
            type: boolean
          in: query
          required: false
        - name: limit
          description: Set the limit on the number of destinations returned. The default is 20.
          schema:
            format: int32
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          in: query
        - name: offset
          description: Set the offset to start at when returning destinations. The default is 0
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          in: query
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationCreateRequest"
            examples:
              Destination Creation Request Example:
                value:
                  name: Postgres
                  workspaceId: 2155ae5a-de39-4808-af6a-16fe7b8b4ed2
                  configuration:
                    airbyte_destination_name: postgres
                    port: 5432
                    schema: public
                    ssl_mode:
                      mode: prefer
                    tunnel_method:
                      tunnel_method: NO_TUNNEL
                    host: localhost
                    database: postgres
                    username: postgres
                    password: test
      tags:
        - Destinations
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
              examples:
                Destination Creation Response Example:
                  value:
                    destinationId: af0c3c67-aa61-419f-8922-95b0bf840e86
          description: Successful operation
        "400":
          description: Invalid data
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: createDestination
      summary: Create a destination
      description: "Creates a destination given a name, workspace id, and a json blob containing the configuration for the source."
  /destinations/{destinationId}:
    get:
      tags:
        - Destinations
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
              examples:
                Destination Get Response Example:
                  value:
                    destinationId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: My Destination
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
                    configuration:
                      conversion_window_days: 14
                      customer_id: "1234567890"
                      start_date: 2023-01-01
                      end_date: 2024-01-01
          description: Get a Destination by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: getDestination
      summary: Get Destination details
    delete:
      tags:
        - Destinations
      responses:
        "204":
          description: The resource was deleted successfully
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: deleteDestination
      summary: Delete a Destination
    patch:
      tags:
        - Destinations
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationPatchRequest"
            examples:
              Destination Update Request Example:
                value:
                  configuration:
                    conversion_window_days: 14
                    customer_id: "1234567890"
                    start_date: 2023-01-01
                    end_date: 2024-01-01
                  name: My Destination
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
              examples:
                Destination Update Response Example:
                  value:
                    destinationId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: running
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
                    configuration:
                      conversion_window_days: 14
                      customer_id: "1234567890"
                      start_date: 2023-01-01
                      end_date: 2024-01-01
          description: Update a Destination
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: patchDestination
      summary: Update a Destination
    put:
      tags:
        - Destinations
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationPutRequest"
            examples:
              Destination Update Request Example:
                value:
                  configuration:
                    conversion_window_days: 14
                    customer_id: "1234567890"
                    start_date: 2023-01-01
                    end_date: 2024-01-01
                  name: My Destination
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
              examples:
                Destination Update Response Example:
                  value:
                    destinationId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: running
                    sourceType: postgres
                    workspaceId: 744cc0ed-7f05-4949-9e60-2a814f90c035
                    configuration:
                      conversion_window_days: 14
                      customer_id: "1234567890"
                      start_date: 2023-01-01
                      end_date: 2024-01-01
          description: Update a Destination and fully overwrite it
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: putDestination
      summary: Update a Destination and fully overwrite it
    parameters:
      - name: destinationId
        schema:
          format: UUID
          type: string
        in: path
        required: true
...
  /workspaces/{workspaceId}:
    parameters:
      - name: workspaceId
        schema:
          format: UUID
          type: string
        in: path
        required: true
    get:
      tags:
        - Workspaces
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
              examples:
                Workspace Get Response Example:
                  value:
                    workspaceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: Acme Company
                    dataResidency: auto
          description: Get a Workspace by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: getWorkspace
      summary: Get Workspace details
    patch:
      tags:
        - Workspaces
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdateRequest"
            examples:
              Workspace Update Request Example:
                value:
                  name: Company Workspace Name
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
              examples:
                Workspace Update Response Example:
                  value:
                    workspaceId: 9924bcd0-99be-453d-ba47-c2c9766f7da5
          description: Successful operation
        "400":
          description: Invalid data
        "403":
          description: Not allowed
      operationId: updateWorkspace
      summary: Update a workspace
    delete:
      tags:
        - Workspaces
      responses:
        "204":
          description: The resource was deleted successfully
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: deleteWorkspace
      summary: Delete a Workspace
...
components:
...
  schemas:
...
    JobResponse:
      title: Root Type for JobResponse
      description: Provides details of a single job.
      required:
        - jobId
        - status
        - jobType
        - startTime
        - connectionId
      type: object
      properties:
        jobId:
          format: int64
          type: integer
        status:
          $ref: "#/components/schemas/JobStatusEnum"
        jobType:
          $ref: "#/components/schemas/JobTypeEnum"
        startTime:
          type: string
        connectionId:
          format: UUID
          type: string
        lastUpdatedAt:
          type: string
        duration:
          description: "Duration of a sync in ISO_8601 format"
          type: string
        bytesSynced:
          format: int64
          type: integer
        rowsSynced:
          format: int64
          type: integer
      example:
        id: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
        status: running
        jobType: sync
        startTime: 2023-03-25T01:30:50Z
        duration: PT8H6M12S
...
    SourcePatchRequest:
      type: object
      properties:
        name:
          type: string
          example: My source
        workspaceId:
          format: uuid
          type: string
        configuration:
          $ref: "#/components/schemas/SourceConfiguration"
        secretId:
          description: Optional secretID obtained through the public API OAuth redirect flow.
          type: string
      x-implements: io.airbyte.api.common.ConfigurableActor
...
    SourcesResponse:
      title: Root Type for SourcesResponse
      description: ""
      required:
        - data
      type: object
      properties:
        previous:
          type: string
        next:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SourceResponse"
      example:
        next: "https://api.airbyte.com/v1/sources?limit=5&offset=10"
        previous: "https://api.airbyte.com/v1/sources?limit=5&offset=0"
        data:
          sourceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
          name: Analytics Team Postgres
          sourceType: postgres
          workspaceId: 871d9b60-11d1-44cb-8c92-c246d53bf87e
...